--- src/sse.rs
@@ -1,1 +1,0 @@
-mod client;
