--- src/sse/client.rs
@@ -1,3 +1,0 @@
-#[cfg(test)]
-
-mod tests {}
